cybexp-master/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.0.7
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
Naked==0.1.31
pkg-resources==0.0.0
pycryptodome==3.6.6
PyJWT==1.6.4
pymongo==3.7.2
pytz==2018.5
PyYAML==3.13
requests==2.20.0
shellescape==3.4.1
urllib3==1.24
